requirements.txt
# CLI
typer>=0.9.0
rich>=13.7.0

# 設定管理
dynaconf>=3.2.4
python-dotenv>=1.0.0

# データ処理
pandas>=2.1.4
openpyxl>=3.1.2
xlwings>=0.30.12; platform_system == "Windows"

# API連携
httpx>=0.26.0
requests>=2.31.0

# 監視・スケジュール機能
watchdog>=3.0.0
apscheduler>=3.10.4

# ロギング・デバッグ
loguru>=0.7.2
icecream>=2.1.3
viztracer>=0.16.1

# 進捗表示
tqdm>=4.66.2

# テスト
pytest>=7.4.3